+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* [bug] exit gracefully when downloaded or restored database backup doesn't exist
 * [bug] 'shelly config list' doesn't quit with error if cloud has no configuration files
 * [improvement] 'shelly backup download' is now aliased to 'shelly backup get'